COMP: fix s/us warning

+++ Core/DataStructures/mitkImage.cpp
@@ -1272,14 +1272,14 @@ void mitk::_ComputeExtremaInItkImage(ItkImageType* itkImage, mitk::Image* mitkIm
 
 const bool mitk::Image::IsValidTimeStep(int t) const
 {
-  return ( ( m_Dimension >= 4 && t <= m_Dimensions[3] && t > 0 ) || (t == 0) ); 
+  return ( ( m_Dimension >= 4 && t <= (int)m_Dimensions[3] && t > 0 ) || (t == 0) ); 
 }
 
 const void mitk::Image::Expand( int timeSteps ) const
 {
   if(timeSteps < 1) itkExceptionMacro(<< "Invalid timestep in Image!");
   if(! IsValidTimeStep( timeSteps-1 ) ) return;
-  if(timeSteps > m_ScalarMin.size() )
+  if(timeSteps > (int)m_ScalarMin.size() )
   {
     m_ScalarMin.resize(timeSteps, itk::NumericTraits<ScalarType>::max());
     m_ScalarMax.resize(timeSteps, itk::NumericTraits<ScalarType>::NonpositiveMin());
